server/requirements.txt
asgiref==3.7.2
Django
tzdata==2023.3
gunicorn
djangorestframework
django-cors-headers
sqlparse==0.2.4
dnspython
psycopg2-binary